implemented @UserData -decorator for decoding user-object from token

+++ src/messages/messages.controller.ts
@@ -9,35 +9,39 @@ import { MessagesService } from './messages.service';
 import { MessageModel } from './message.model';
 import { JwtAuthGuard } from '../auth/guards/jwt.guard';
 import { UserData } from '../infrastructure/decorators/user-data.decorator';
-import { ACCESS_DENIED, NOT_FOUND_ERROR } from '../infrastructure/constants';
-import { Role, UserModel } from '../auth/user.model';
+import { ACCESS_DENIED, BAD_REQUEST, NOT_FOUND_ERROR } from '../infrastructure/constants';
+import { DecodedUser } from '../infrastructure/interfaces/decoded-user.interface';
+import { Role } from '../infrastructure/enums/roles.enum';
 
 @Controller('messages')
 @UseGuards(JwtAuthGuard)
 export class MessagesController {
     constructor(private readonly messagesService: MessagesService) { }
 
-    // @Get()
-    // @HttpCode(200)
-    // @ApiOperation({ summary: 'Get all messages' })
-    // async findAll(@UserData() userFromRequest: { user: UserModel }): Promise<MessageModel[]> {
-    //     if (!userFromRequest.user.roles.includes(Role.Admin)) {
-    //         throw new HttpException(ACCESS_DENIED, HttpStatus.FORBIDDEN);
-    //     }
-    //     return await this.messagesService.findAll();
-    // }
-
     @Get()
     @HttpCode(200)
-    @ApiOperation({ summary: 'Get all messages' })
-    async findAll(): Promise<MessageModel[]> {
+    @ApiOperation({ summary: 'Get all messages, allows access for users with admin role' })
+    async findAll(@UserData() decodedUser: DecodedUser): Promise<MessageModel[]> {
+        if (!decodedUser) {
+            throw new HttpException(BAD_REQUEST, HttpStatus.BAD_REQUEST);
+        }
+        if (!decodedUser.user.roles.includes(Role.Admin)) {
+            throw new HttpException(ACCESS_DENIED, HttpStatus.UNAUTHORIZED);
+        }
         return await this.messagesService.findAll();
     }
 
     @Get(':id')
     @HttpCode(200)
-    @ApiOperation({ summary: 'Get message by id' })
-    async findById(@Param('id') id: string): Promise<MessageModel> {
+    @ApiOperation({ summary: 'Get message by id, allows access for users with admin role' })
+    async findById(@UserData() decodedUser: DecodedUser, @Param('id') id: string): Promise<MessageModel> {
+        if (!decodedUser) {
+            throw new HttpException(BAD_REQUEST, HttpStatus.BAD_REQUEST);
+        }
+        if (!decodedUser.user.roles.includes(Role.Admin)) {
+            throw new HttpException(ACCESS_DENIED, HttpStatus.UNAUTHORIZED);
+        }
+
         const message = await this.messagesService.findById(id);
         if (!message) {
             throw new HttpException(NOT_FOUND_ERROR, HttpStatus.NOT_FOUND);
@@ -48,23 +52,38 @@ export class MessagesController {
     @Post()
     @HttpCode(201)
     @UsePipes(new ValidationPipe())
-    @ApiOperation({ summary: 'Create message' })
-    async create(@Body() createMessageDto: CreateMessageDto) {
+    @ApiOperation({ summary: 'Create message, allows access for registered users' })
+    async create(@UserData() decodedUser: DecodedUser, @Body() createMessageDto: CreateMessageDto) {
+        if (!decodedUser) {
+            throw new HttpException(ACCESS_DENIED, HttpStatus.UNAUTHORIZED);
+        }
         return await this.messagesService.create(createMessageDto);
     }
 
     @Delete(':id')
     @HttpCode(200)
-    @ApiOperation({ summary: 'Delete message by id' })
-    async deleteById(@Param('id') id: string) {
+    @ApiOperation({ summary: 'Delete message by id, allows access for users with admin role' })
+    async deleteById(@UserData() decodedUser: DecodedUser, @Param('id') id: string) {
+        if (!decodedUser) {
+            throw new HttpException(BAD_REQUEST, HttpStatus.BAD_REQUEST);
+        }
+        if (!decodedUser.user.roles.includes(Role.Admin)) {
+            throw new HttpException(ACCESS_DENIED, HttpStatus.UNAUTHORIZED);
+        }
         return await this.messagesService.deleteById(id);
     }
 
     @Put(':id')
     @HttpCode(200)
     @UsePipes(new ValidationPipe())
-    @ApiOperation({ summary: 'Replace old message by new instance, using id' })
-    async updateById(@Param('id') id: string, @Body() updateMessageDto: CreateMessageDto) {
+    @ApiOperation({ summary: 'Replace old message by new instance, using id, allows access for users with admin role' })
+    async updateById(@UserData() decodedUser: DecodedUser, @Param('id') id: string, @Body() updateMessageDto: CreateMessageDto) {
+        if (!decodedUser) {
+            throw new HttpException(BAD_REQUEST, HttpStatus.BAD_REQUEST);
+        }
+        if (!decodedUser.user.roles.includes(Role.Admin)) {
+            throw new HttpException(ACCESS_DENIED, HttpStatus.UNAUTHORIZED);
+        }
         return await this.messagesService.updateById(id, updateMessageDto);
     }
 }
